Add FeatureTest (#115)
* Remove unused method in esb test

* Refactor UserJourney into multiple feature tests

- This speeds up test execution by allowing them to run in paralell now that they're in different files

* remove integration tag from PdfIntegrationTest
