+++ src/test/java/org/codeforamerica/shiba/pages/features/UserJourneyPageTest.java
@@ -1,0 +1,178 @@
+package org.codeforamerica.shiba.pages.features;
+
+import org.apache.pdfbox.pdmodel.PDDocument;
+import org.apache.pdfbox.pdmodel.interactive.form.PDAcroForm;
+import org.codeforamerica.shiba.pages.SuccessPage;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.CsvSource;
+import org.openqa.selenium.By;
+
+import java.io.File;
+import java.io.IOException;
+import java.util.Arrays;
+import java.util.List;
+import java.util.concurrent.TimeUnit;
+
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.awaitility.Awaitility.await;
+import static org.codeforamerica.shiba.pages.YesNoAnswer.NO;
+import static org.codeforamerica.shiba.pages.YesNoAnswer.YES;
+
+public class UserJourneyPageTest extends FeatureTest {
+    @Test
+    void intercomButtonIsPresent() {
+        driver.manage().timeouts().implicitlyWait(5, TimeUnit.SECONDS);
+        assertThat(driver.findElementById("intercom-frame")).isNotNull();
+    }
+
+    @Test
+    void userCanCompleteTheNonExpeditedHouseholdFlow() {
+        nonExpeditedFlowToSuccessPage(true, true, smartyStreetClient);
+    }
+
+    @Test
+    void userCanCompleteTheNonExpeditedFlowWithNoEmployment() {
+        nonExpeditedFlowToSuccessPage(false, false, smartyStreetClient);
+    }
+
+    @Test
+    void userCanCompleteTheNonExpeditedHouseholdFlowWithNoEmployment() {
+        nonExpeditedFlowToSuccessPage(true, false, smartyStreetClient);
+    }
+
+    @ParameterizedTest
+    @CsvSource(value = {
+            "123, 1233, A caseworker will contact you within 5-7 days to review your application.",
+            "1, 1, A caseworker will contact you within 3 days to review your application."
+    })
+    void userCanCompleteTheExpeditedFlow(String moneyMadeLast30Days, String liquidAssets, String expeditedServiceDetermination) {
+        completeFlowFromLandingPageThroughReviewInfo(List.of(PROGRAM_SNAP, PROGRAM_CCAP), smartyStreetClient);
+        testPage.clickLink("Submit application now with only the above information.");
+        testPage.clickLink("Yes, I want to see if I qualify");
+
+        testPage.enter("addHouseholdMembers", NO.getDisplayValue());
+        testPage.enter("moneyMadeLast30Days", moneyMadeLast30Days);
+
+        testPage.clickContinue();
+        testPage.enter("haveSavings", YES.getDisplayValue());
+
+        testPage.enter("liquidAssets", liquidAssets);
+
+        testPage.clickContinue();
+        testPage.enter("payRentOrMortgage", YES.getDisplayValue());
+
+        testPage.enter("homeExpensesAmount", "333");
+        testPage.clickContinue();
+
+        testPage.enter("payForUtilities", "Cooling");
+        testPage.clickContinue();
+
+        testPage.enter("migrantOrSeasonalFarmWorker", NO.getDisplayValue());
+
+        assertThat(driver.findElement(By.tagName("p")).getText()).contains(expeditedServiceDetermination);
+
+        testPage.clickButton("Finish application");
+        assertThat(testPage.getTitle()).isEqualTo("Legal Stuff");
+        assertThat(driver.findElement(By.id("ccap-legal"))).isNotNull();
+    }
+
+    @ParameterizedTest
+    @CsvSource(value = {
+            "123, 1233, A caseworker will contact you within 5-7 days to review your application.",
+            "1, 1, A caseworker will contact you within 3 days to review your application."
+    })
+    void userCanCompleteTheExpeditedFlowWithHousehold(String moneyMadeLast30Days, String liquidAssets, String expeditedServiceDetermination) {
+        completeFlowFromLandingPageThroughReviewInfo(List.of(PROGRAM_SNAP, PROGRAM_CCAP), smartyStreetClient);
+        testPage.clickLink("Submit application now with only the above information.");
+        testPage.clickLink("Yes, I want to see if I qualify");
+
+        testPage.enter("addHouseholdMembers", YES.getDisplayValue());
+        testPage.enter("moneyMadeLast30Days", moneyMadeLast30Days);
+
+        testPage.clickContinue();
+        testPage.enter("haveSavings", YES.getDisplayValue());
+
+        testPage.enter("liquidAssets", liquidAssets);
+
+        testPage.clickContinue();
+        testPage.enter("payRentOrMortgage", YES.getDisplayValue());
+
+        testPage.enter("homeExpensesAmount", "333");
+        testPage.clickContinue();
+
+        testPage.enter("payForUtilities", "Cooling");
+        testPage.clickContinue();
+
+        testPage.enter("migrantOrSeasonalFarmWorker", NO.getDisplayValue());
+
+        assertThat(driver.findElement(By.tagName("p")).getText()).contains(expeditedServiceDetermination);
+
+        testPage.clickButton("Finish application");
+        assertThat(testPage.getTitle()).isEqualTo("Legal Stuff");
+        assertThat(driver.findElement(By.id("ccap-legal"))).isNotNull();
+    }
+
+    @Test
+    void partialFlow() throws IOException {
+        getToDocumentUploadScreen();
+        completeDocumentUploadFlow();
+
+        SuccessPage successPage = new SuccessPage(driver);
+        successPage.downloadPdfs();
+        await().until(() -> {
+            File[] listFiles = path.toFile().listFiles();
+            return Arrays.stream(listFiles).anyMatch(file -> file.getName().contains("_MNB_") && file.getName().endsWith(".pdf"));
+        });
+
+        File pdfFile = Arrays.stream(path.toFile().listFiles()).findFirst().orElseThrow();
+        PDAcroForm acroForm = PDDocument.load(pdfFile).getDocumentCatalog().getAcroForm();
+        assertThat(acroForm.getField("APPLICANT_WRITTEN_LANGUAGE_PREFERENCE").getValueAsString())
+                .isEqualTo("ENGLISH");
+        assertThat(acroForm.getField("APPLICANT_WRITTEN_LANGUAGE_PREFERENCE").getValueAsString())
+                .isEqualTo("ENGLISH");
+        assertThat(acroForm.getField("NEED_INTERPRETER").getValueAsString())
+                .isEqualTo("Yes");
+    }
+
+    @Test
+    void deletingUploadedFileShouldLoadDocumentUploadScreenUponConfirmDeletion() {
+        getToDocumentUploadScreen();
+        uploadDefaultFile();
+        uploadDefaultFile();
+
+        testPage.clickLink("delete");
+
+        assertThat(testPage.getTitle()).isEqualTo("Delete a file");
+        testPage.clickButton("Yes, delete the file");
+
+        assertThat(testPage.getTitle()).isEqualTo("Upload Documents");
+    }
+
+    @Test
+    void shouldHandleDeletionOfLastHouseholdMember() {
+        completeFlowFromLandingPageThroughReviewInfo(List.of(PROGRAM_CCAP), smartyStreetClient);
+        testPage.clickLink("This looks correct");
+        testPage.enter("addHouseholdMembers", YES.getDisplayValue());
+        testPage.clickContinue();
+        fillOutHousemateInfo(PROGRAM_EA);
+        testPage.clickContinue();
+        testPage.clickButtonLink("delete");
+        testPage.clickButton("Yes, remove them");
+        testPage.goBack();
+
+        assertThat(driver.getTitle()).isEqualTo("Review info");
+    }
+
+    @Test
+    void shouldValidateContactInfoEmailEvenIfEmailNotSelected() {
+        completeFlowFromLandingPageThroughContactInfo(List.of(PROGRAM_CCAP));
+        testPage.enter("phoneNumber", "7234567890");
+        testPage.enter("email", "email.com");
+        testPage.enter("phoneOrEmail", "Text me");
+        testPage.clickContinue();
+
+        assertThat(driver.getTitle()).isEqualTo("Contact Info");
+    }
+
+}
